[REFACTOR] fixture 패키지 분리
