+++ src/test/java/com/soptie/server/doll/service/DollServiceImplTest.java
@@ -4,14 +4,11 @@ import com.soptie.server.doll.dto.DollImageResponse;
 import com.soptie.server.doll.entity.Doll;
 import com.soptie.server.doll.entity.DollType;
 import com.soptie.server.doll.repository.DollRepository;
-import com.soptie.server.support.DollFixture;
+import com.soptie.server.support.fixture.DollFixture;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.EnumSource;
-import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
