+++ src/test/java/com/soptie/server/member/service/MemberServiceImplTest.java
@@ -13,10 +13,10 @@ import com.soptie.server.member.repository.MemberRepository;
 import com.soptie.server.memberDoll.entity.MemberDoll;
 import com.soptie.server.memberDoll.service.MemberDollServiceImpl;
 import com.soptie.server.memberRoutine.service.MemberDailyRoutineServiceImpl;
-import com.soptie.server.support.ConversationFixture;
-import com.soptie.server.support.DollFixture;
-import com.soptie.server.support.MemberDollFixture;
-import com.soptie.server.support.MemberFixture;
+import com.soptie.server.support.fixture.ConversationFixture;
+import com.soptie.server.support.fixture.DollFixture;
+import com.soptie.server.support.fixture.MemberDollFixture;
+import com.soptie.server.support.fixture.MemberFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
