+++ src/test/java/com/soptie/server/memberRoutine/repository/CompletedMemberDailyRoutineRepositoryTest.java
@@ -17,10 +17,10 @@ import com.soptie.server.routine.entity.daily.DailyRoutine;
 import com.soptie.server.routine.entity.daily.DailyTheme;
 import com.soptie.server.routine.repository.daily.routine.DailyRoutineRepository;
 import com.soptie.server.routine.repository.daily.theme.DailyThemeRepository;
-import com.soptie.server.support.CompletedMemberDailyRoutineFixture;
-import com.soptie.server.support.DailyRoutineFixture;
-import com.soptie.server.support.DailyThemeFixture;
-import com.soptie.server.support.MemberFixture;
+import com.soptie.server.support.fixture.CompletedMemberDailyRoutineFixture;
+import com.soptie.server.support.fixture.DailyRoutineFixture;
+import com.soptie.server.support.fixture.DailyThemeFixture;
+import com.soptie.server.support.fixture.MemberFixture;
 import com.soptie.server.support.RepositoryTest;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
