+++ src/test/java/com/soptie/server/memberRoutine/service/MemberDailyRoutineServiceImplTest.java
@@ -28,10 +28,10 @@ import com.soptie.server.routine.entity.daily.DailyRoutine;
 import com.soptie.server.routine.entity.daily.DailyTheme;
 import com.soptie.server.routine.exception.RoutineException;
 import com.soptie.server.routine.repository.daily.routine.DailyRoutineRepository;
-import com.soptie.server.support.DailyRoutineFixture;
-import com.soptie.server.support.DailyThemeFixture;
-import com.soptie.server.support.MemberDailyRoutineFixture;
-import com.soptie.server.support.MemberFixture;
+import com.soptie.server.support.fixture.DailyRoutineFixture;
+import com.soptie.server.support.fixture.DailyThemeFixture;
+import com.soptie.server.support.fixture.MemberDailyRoutineFixture;
+import com.soptie.server.support.fixture.MemberFixture;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @SuppressWarnings("NonAsciiCharacters")
