+++ src/test/java/com/soptie/server/memberRoutine/service/MemberHappinessRoutineServiceImplTest.java
@@ -9,8 +9,8 @@ import com.soptie.server.memberRoutine.entity.happiness.MemberHappinessRoutine;
 import com.soptie.server.memberRoutine.repository.MemberHappinessRoutineRepository;
 import com.soptie.server.routine.entity.happiness.HappinessSubRoutine;
 import com.soptie.server.routine.repository.happiness.routine.HappinessSubRoutineRepository;
-import com.soptie.server.support.MemberFixture;
-import com.soptie.server.support.MemberHappinessRoutineFixture;
+import com.soptie.server.support.fixture.MemberFixture;
+import com.soptie.server.support.fixture.MemberHappinessRoutineFixture;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
