+++ src/test/java/com/soptie/server/routine/service/HappinessRoutineServiceImplTest.java
@@ -6,8 +6,8 @@ import com.soptie.server.routine.entity.happiness.HappinessRoutine;
 import com.soptie.server.routine.entity.happiness.HappinessTheme;
 import com.soptie.server.routine.repository.happiness.routine.HappinessRoutineRepository;
 import com.soptie.server.routine.repository.happiness.theme.HappinessThemeRepository;
-import com.soptie.server.support.HappinessRoutineFixture;
-import com.soptie.server.support.HappinessThemeFixture;
+import com.soptie.server.support.fixture.HappinessRoutineFixture;
+import com.soptie.server.support.fixture.HappinessThemeFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
